requirements.txt
pandas
google-cloud-bigquery
openai
streamlit
requests
